+++ fcrepo-kernel-impl/src/main/java/org/fcrepo/kernel/impl/operations/RdfSourceOperationFactoryImpl.java
@@ -31,7 +31,8 @@ public class RdfSourceOperationFactoryImpl implements RdfSourceOperationFactory 
 
     @Override
     public RdfSourceOperationBuilder createBuilder(final String rescId, final String interactionModel) {
-        return new CreateRdfSourceOperationBuilder(rescId, interactionModel);
+        // TODO: Not sure what we need the interactionModel for here?
+        return new CreateRdfSourceOperationBuilder(rescId);
     }
 
     @Override
